# [IMP] mrp: improve _compute_quantities_dict performance