We can iterate through boms instead of products to avoid computing the same quantities multiple times, for example in the case of variants which don't have different boms from the template. This works only for boms which don't have any line with "apply on variant". For each of these bom kits we can compute the quantities once and apply them at the end for all relevant products.

Benchmarks:

Reading manufactured products in the Products view. Among these, 3 products have bom kits without any "apply on variant" line and 3 products have bom kits with only "apply on variant" lines. The bom kits have 1000 components each and the products have 1000 variants each.

| |before|	after|
|--|--|--|
|only apply|		6.8 minutes|	6.9 minutes|	
|only without|		2.5 minutes|	1.54 seconds|	
|both|			9.3 minutes|	6.8 minutes|

As expected, this PR only improves the performance of boms without any "apply on variant" line. However, it doesn't significantly impact performance on those boms with "apply on variant" lines.

For reference, here are more benchmarks taken only on boms without any "apply on variant" lines. Note that this was on a slightly different patch, but the times shouldn't be significantly different.

(Opening the form view of a product with a BoM Kit)

|			| |2k components|	1k components|	500 components|	100 components|
|--|--|--|--|--|--|
|1k variants	|before	|1.6 min|	48.37 s|	28.24 s|	9.33 s|
|		|after	|883 ms|		631 ms|		477 ms|		411 ms|
|500 variants	|before	|48.17 s|	26.65 s|	15.09 s|	4.73 s|
|		|after	|694 ms|		495 ms|		416 ms|		333 ms|
|250 variants	|before	|26.35 s|	15.32 s|	9.11 s|		2.59 s|
|		|after	|635 ms|		467 ms|		383 ms|		295 ms|	
|125 variants	|before	|14.63 s|	9.12 s|		4.88 s|		1.42 s|
|		|after	|649 ms|		438 ms|		422 ms|		295 ms|



--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr